Juggle more namespaces
