+++ DbgCensus.EventStream.EventHandlers/Extensions/IServiceCollectionExtensions.cs
@@ -4,8 +4,8 @@ using DbgCensus.EventStream.Abstractions.Objects.Events.Worlds;
 using DbgCensus.EventStream.EventHandlers.Abstractions;
 using DbgCensus.EventStream.EventHandlers.Objects.Event;
 using DbgCensus.EventStream.Extensions;
-using DbgCensus.EventStream.Objects.Characters;
-using DbgCensus.EventStream.Objects.Worlds;
+using DbgCensus.EventStream.Objects.Events.Characters;
+using DbgCensus.EventStream.Objects.Events.Worlds;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Logging;
