+++ DbgCensus.EventStream/Objects/Events/Characters/SkillAdded.cs
@@ -2,25 +2,23 @@
 using DbgCensus.EventStream.Abstractions.Objects.Events.Characters;
 using System;
 
-namespace DbgCensus.EventStream.Objects.Characters;
+namespace DbgCensus.EventStream.Objects.Events.Characters;
 
 /// <summary>
-/// Initializes a new instance of the <see cref="PlayerFacilityCapture"/> record.
+/// Initializes a new instance of the <see cref="SkillAdded"/> record.
 /// </summary>
-/// <param name="CharacterID">The ID of the character for which this event was generated.</param>
+/// <param name="CharacterID">The ID of the character for which the event was generated.</param>
 /// <param name="EventName">The name of the event.</param>
-/// <param name="FacilityID">The ID of the captured facility.</param>
-/// <param name="OutfitID">The ID of the outfit that the <paramref name="CharacterID"/> belongs to.</param>
+/// <param name="SkillID">The ID of the unlocked skill.</param>
 /// <param name="Timestamp">The time at which the event occured.</param>
 /// <param name="WorldID">The world on which the event occured.</param>
 /// <param name="ZoneID">The zone on which the event occured.</param>
-public record PlayerFacilityCapture
+public record SkillAdded
 (
     ulong CharacterID,
     string EventName,
-    uint FacilityID,
-    ulong OutfitID,
+    uint SkillID,
     DateTimeOffset Timestamp,
     WorldDefinition WorldID,
     ZoneID ZoneID
-) : IPlayerFacilityCapture;
+) : ISkillAdded;
